Bug 1290644 Remove VS2010 workaround in dom/media/. r=ayang

+++ dom/media/encoder/fmp4_muxer/ISOMediaBoxes.cpp
@@ -1401,8 +1401,7 @@ FullBox::FullBox(const nsACString& aType, uint8_t aVersion, uint32_t aFlags,
                  ISOControl* aControl)
   : Box(aType, aControl)
 {
-  // Cast to uint64_t due to VC2010  bug.
-  std::bitset<24> tmp_flags((uint64_t)aFlags);
+  std::bitset<24> tmp_flags(aFlags);
   version = aVersion;
   flags = tmp_flags;
   size += sizeof(version) + flags.size() / CHAR_BIT;
